djangoapp/requirements.txt
asgiref==3.2.3
astroid==2.3.3
Babel==2.8.0
colorama==0.4.3
Django==3.0.3
django-appconf==1.0.3
django-crispy-forms==1.8.1
django-grappelli==2.14.1
django-jet==1.0.8
django-phonenumber-field==4.0.0
django-select2==7.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.11.4
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
